load_testing/requirements.txt
faker
reportlab
google-cloud-storage